Refactory main.dart
